Add files via upload
I remodeled the CSS, because I was trying to positioning the tags with position and not with margin, because I thought it will be better, but I was wrong, I had to remake basically all the CSS code, however I learned with my fail and now I positioned the thing with margin, I searched how to center a text in a HR tag, and I did it. This commit is a victory for me because I surpassed the trouble and now I will keep going, praying to finish this in the time. Thanks a lot for the challenge, I was needing this to upgrade more and more. Now let's go to Javascript part of Desktop, I don't have so much experience with API, but I know I can do it.